divide Faucet into a container and a component
